requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
Babel==2.9.1
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
colorama==0.4.4
dill==0.3.4
editdistance==0.5.3
filelock==3.0.12
flatbuffers==1.12
future==0.18.2
gast==0.4.0
gin-config==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.5
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grpcio==1.39.0
h5py==3.1.0
huggingface-hub==0.0.12
idna==3.2
iniconfig==1.1.1
joblib==1.0.1
keras==2.6.0
Keras-Preprocessing==1.1.2
Levenshtein==0.13.0
Markdown==3.3.4
mesh-tensorflow==0.1.19
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.2
pluggy==0.13.1
portalocker==2.3.0
promise==2.3
protobuf==3.17.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rouge-score==0.0.4
rsa==4.7.2
sacrebleu==2.0.0
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
sentencepiece==0.1.96
seqio==0.0.6
six==1.15.0
t5==0.9.2
tabulate==0.8.9
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-datasets==4.4.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.2.0
tensorflow-text==2.6.0
termcolor==1.1.0
tfds-nightly==4.4.0.dev202108200109
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
tqdm==4.62.1
transformers==4.9.2
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
